Move growatt_server sensor definitions
