+++ homeassistant/components/growatt_server/sensor/inverter.py
