+++ homeassistant/components/growatt_server/sensor/sensor_entity_description.py
@@ -15,7 +15,6 @@ class GrowattRequiredKeysMixin:
 
 
 @dataclass(frozen=True)
-# pylint: disable-next=hass-enforce-class-module
 class GrowattSensorEntityDescription(SensorEntityDescription, GrowattRequiredKeysMixin):
     """Describes Growatt sensor entity."""
