+++ homeassistant/components/growatt_server/sensor/storage.py
