+++ homeassistant/components/growatt_server/sensor/tlx.py
