+++ homeassistant/components/growatt_server/sensor/total.py
